tui: Add struct graph_time_stat and use it in uftrace_graph_node
This patch introduces struct graph_time_stat that maintains the total
execution time of graph node and it replaces uint64_t time as follows.

Before:
  struct uftrace_graph_node {
              ...
          uint64_t time;
   	  uint64_t child_time;
              ...
  };

After:
  struct graph_time_stat {
          uint64_t sum;
  };
  struct uftrace_graph_node {
              ...
          struct graph_time_stat total_time;
   	  uint64_t child_time;
              ...
  };

This patch applies the related changes only for this change and nothing
more unrelated.

The struct graph_time_stat currently has only sum but it will be
expanded with min and max in the follow up patches.

Signed-off-by: Lee Dong Wook <sh95119@gmail.com>
Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>